move conversion functions into seperate file and set up check set suite

+++ thermometer/src/conversions.h
@@ -1,4 +1,4 @@
-float voltage_from_pin(float pinValue);
+float voltage_from_pin(int pinValue);
 float resistance_from_voltage(float voltage);
 float celsius_from_kelvin(float tempKelvin);
 float fahrenheit_from_celsius(float tempCelsius);
